@@ -10,41 +10,39 @@
   },
   "browserslist": "defaults, not ie <= 11",
   "dependencies": {
-    "@algolia/autocomplete-core": "^1.7.3",
-    "@algolia/autocomplete-preset-algolia": "^1.7.3",
-    "@headlessui/react": "^1.7.11",
-    "@mdx-js/loader": "^2.1.5",
-    "@mdx-js/react": "^2.1.5",
-    "@next/mdx": "^13.0.3",
-    "@sindresorhus/slugify": "^2.1.1",
-    "@tailwindcss/typography": "^0.5.8",
-    "acorn": "^8.8.1",
-    "algoliasearch": "^4.14.2",
-    "autoprefixer": "^10.4.7",
-    "clsx": "^1.2.0",
-    "focus-visible": "^5.2.0",
-    "framer-motion": "7.8.1",
-    "mdast-util-to-string": "^3.1.0",
-    "mdx-annotations": "^0.1.1",
-    "multiparty": "^4.2.3",
-    "next": "13.0.2",
-    "postcss-focus-visible": "^6.0.4",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "recma-nextjs-static-props": "^1.0.0",
-    "rehype-mdx-title": "^2.0.0",
-    "remark-gfm": "^3.0.1",
-    "shiki": "^0.13.0",
-    "tailwindcss": "^3.2.4",
-    "unist-util-visit": "^4.1.1",
-    "zustand": "^4.3.2"
+    "@headlessui/react": "^2.2.0",
+    "@mdx-js/loader": "^3.1.0",
+    "@mdx-js/react": "^3.1.0",
+    "@next/mdx": "^15.1.3",
+    "@sindresorhus/slugify": "^2.2.1",
+    "@tailwindcss/typography": "^0.5.15",
+    "@types/mdx": "^2.0.13",
+    "acorn": "^8.14.0",
+    "autoprefixer": "^10.4.20",
+    "clsx": "^2.1.1",
+    "framer-motion": "^11.13.5",
+    "mdast-util-to-string": "^4.0.0",
+    "mdx-annotations": "^0.1.4",
+    "next": "^15.1.3",
+    "postcss": "^8.4.49",
+    "postcss-focus-visible": "^10.0.1",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "rehype-mdx-title": "^3.0.1",
+    "remark-gfm": "^4.0.0",
+    "shiki": "^1.24.2",
+    "tailwindcss": "^3.4.17",
+    "unist-util-visit": "^5.0.0",
+    "zustand": "^5.0.2"
   },
   "devDependencies": {
-    "@types/node": "18.15.1",
-    "eslint": "8.26.0",
-    "eslint-config-next": "13.0.2",
-    "prettier": "^2.7.1",
-    "prettier-plugin-tailwindcss": "^0.1.13",
-    "typescript": "4.9.5"
+    "@types/node": "^22.10.2",
+    "@types/react": "^19.0.1",
+    "@types/react-dom": "^19.0.2",
+    "eslint": "^9.17.0",
+    "eslint-config-next": "^15.1.3",
+    "prettier": "^3.4.2",
+    "prettier-plugin-tailwindcss": "^0.6.9",
+    "typescript": "^5.7.2"
   }
 }
