@@ -4,21 +4,18 @@ import { rehypePlugins } from './mdx/rehype.mjs'
 import { recmaPlugins } from './mdx/recma.mjs'
 
 const withMDX = nextMDX({
+  extension: /\.mdx?$/,
   options: {
     remarkPlugins,
     rehypePlugins,
     recmaPlugins,
-    providerImportSource: '@mdx-js/react',
   },
 })
 
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   reactStrictMode: true,
-  pageExtensions: ['js', 'jsx', 'ts', 'tsx', 'mdx'],
-  experimental: {
-    scrollRestoration: true,
-  },
+  pageExtensions: ['js', 'jsx', 'ts', 'tsx', 'md', 'mdx'],
   async rewrites() {
     return [
       {
